+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 
 ### Removes
 
-- Remove the `outputPath` parameter. Templates are generated in the the webpack output path. Use the `filename` to customize the output directory.
+- Remove the `outputPath` parameter. Templates are generated in the webpack output path. Use the `filename` to customize the output directory.
 - Remove the `fileExtension` parameter.
 - Remove the `fs-extra` package in favor of the file system exposed by the webpack compiler. It use the `graceful-fs` package.
 - Remove the utils and move useful functions into the main file